move test to better home
