+++ drools-model/drools-model-codegen/src/test/java/org/drools/model/codegen/execmodel/GlobalTest.java
@@ -19,6 +19,7 @@
 package org.drools.model.codegen.execmodel;
 
 import java.lang.reflect.Method;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -94,6 +95,32 @@ public class GlobalTest extends BaseModelTest {
         ksession.fireAllRules();
 
         assertThat(result.getValue()).isEqualTo("Mark is 37");
+    }
+
+    @ParameterizedTest
+    @MethodSource("parameters")
+    public void testGlobalReferencedInConstraintFromDifferentPackage(RUN_TYPE runType) {
+        // A global declared in one DRL package must be resolvable from a pattern constraint compiled from
+        // a different package in the same KieBase (globals are KieBase-wide, matching the classic compiler)
+        // and usable at runtime. Here 'threshold' is declared in com.test.globals and referenced from a
+        // Person constraint in com.test.rules, with no wildcard import.
+        String globalDrl =
+                "package com.test.globals;\n" +
+                "global java.math.BigDecimal threshold;\n";
+        String ruleDrl =
+                "package com.test.rules;\n" +
+                "import " + Person.class.getCanonicalName() + ";\n" +
+                "rule R when\n" +
+                "  Person( money.compareTo(threshold) < 0 )\n" +
+                "then end";
+
+        KieSession ksession = getKieSession(runType, globalDrl, ruleDrl);
+        ksession.setGlobal("threshold", new BigDecimal("100"));
+
+        ksession.insert(new Person("Poor", new BigDecimal("50")));   // 50 < 100 -> matches
+        ksession.insert(new Person("Rich", new BigDecimal("150")));  // 150 < 100 -> no match
+
+        assertThat(ksession.fireAllRules()).isEqualTo(1);
     }
 
     public static class Functions {
